myenv/requirements.txt
API Testing
Test that you can get a list of products from the API and that the response contains the expected fields.
Test that you can log in with valid credentials and that the response contains an authentication token.
Test that you cannot log in with invalid credentials and that the response contains an error message.
Test that you can add a product to your cart using the API and that the response indicates that the product was added successfully.
Test that you cannot add a product to your cart if you are not logged in and that the response contains an error message.

Automated Visual Testing
Test that the login page looks correct and contains the expected elements.
Test that the inventory page looks correct and contains the expected elements.
Test that the cart page looks correct and contains the expected elements.
Test that the checkout page looks correct and contains the expected elements.
Test that the confirmation page looks correct and contains the expected elements.

BDD Testing
Test that you can log in with valid credentials.
Test that you cannot log in with invalid credentials.
Test that you can add a product to your cart.
Test that you can remove a product from your cart.
Test that you can complete a purchase and receive a confirmation.

Selenium WebDriver Testing
Test that you can navigate to the login page.
Test that you can log in with valid credentials.
Test that you cannot log in with invalid credentials.
Test that you can add a product to your cart.
Test that you can complete a purchase and receive a confirmation.

Robot Framework Testing
Test that you can navigate to the login page and see the expected elements.
Test that you can log in with valid credentials and see the expected page.
Test that you cannot log in with invalid credentials and see the expected error message.
Test that you can add a product to your cart and see the expected message.
Test that you can complete a purchase and receive a confirmation.